requirements.txt
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
PyGithub==1.58.2
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
bibtexparser==1.4.0
watchdog==3.0.0
click==8.1.3
rich==13.4.2
pathlib==1.0.1
